requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
bcrypt==4.0.1
blinker==1.8.2
bson==0.5.10
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.100.0
fastapi-cli==0.0.4
fastapi-jwt==0.3.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.0.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
